Write-Box has Justification!
Ok, so this is small, but soooo obvious. I modified the write-right and
write-center to allow specifying smaller than screenwidth (or larger,
since I don't validate the amount) numbers. Now, write-box has switches
for right and center justification within the box. (And left, but that's
the default [and old] behavior).